Added Client Select Function

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -59,7 +59,7 @@ public class DishController {
     }
 
     @GetMapping("/{id}")
-    public Result<DishVO> getById(@PathVariable Integer id){
+    public Result<DishVO> getById(@PathVariable Long id){
         DishVO dishVO = dishService.getById(id);
         return Result.success(dishVO);
     }
